requirements.txt
# Core dependencies
PyQt6>=6.5.0
reportlab>=4.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
Pillow>=10.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Optional - for web version
streamlit>=1.28.0
